\name{rnewton-package}
\alias{rnewton-package}
\alias{rnewton}
\docType{package}
\title{\packageTitle{rnewton}}
\description{\packageDescription{rnewton}}
\section{Package Content}{\packageIndices{rnewton}}
\author{\packageAuthor{rnewton}}
\section{Maintainer}{\packageMaintainer{rnewton}}
\keyword{package}
